created dto

+++ src/main/java/by/itra/pikachy/api/entity/Post.java
@@ -18,7 +18,7 @@ public class Post {
     @Column(name = "description")
     private String description;
     @Column(name = "mark")
-    private byte mark;
+    private int mark;
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
